feat: update double type

+++ android/src/main/java/com/reactnativearviewer/ArViewerView.kt
@@ -582,14 +582,14 @@ class ArViewerView @JvmOverloads constructor(
   /**
    * Set min max scale for model
    */
-  fun setMinScale(min: Float = 0.1) {
+  fun setMinScale(min: Double = 0.1) {
     if (modelNode == null) return
     modelNode!!.scaleController.setMinScale(min)
   }
   /**
    * Set min max scale for model
    */
-  fun setMaxScale(max: Float = 5) {
+  fun setMaxScale(max: Double = 5) {
     if (modelNode == null) return
     modelNode!!.scaleController.setMaxScale(max)
   }
